+++ packages/cozy-client/src/index.js
@@ -1,6 +1,6 @@
 export { default } from './CozyClient'
 export { default as CozyProvider } from './Provider'
 export { default as connect } from './connect'
-export { default as withMutate } from './withMutate'
+export { default as withMutation } from './withMutation'
 export { all, find } from './Query'
 export { create, update, destroy } from './Mutation'
